inventory_management/requirements.txt
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.28.0
click==8.1.7
databases==0.8.0
exceptiongroup==1.1.3
fastapi==0.103.2
greenlet==3.0.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2==2.9.9
pydantic==1.10.13
pydantic-core==2.10.1
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
typing-extensions==4.8.0
uvicorn==0.23.2
zipp==3.17.0

